add the blind_box

+++ models/common.go
@@ -4,21 +4,21 @@ import "errors"
 
 type BindCFXWithDiscord struct {
 	BaseModel
-	DiscordId string `gorm:"type:varchar(256)" json:"discord_id" binding:"required"`
+	DiscordId  string `gorm:"type:varchar(256)" json:"discord_id" binding:"required"`
 	CFXAddress string `gorm:"type:varchar(256)" json:"cfx_address" binding:"required"`
 }
 
 type BindCFXWithDoDo struct {
 	BaseModel
-	DoDoId string `gorm:"type:varchar(256)" json:"do_do_id" binding:"required"`
+	DoDoId     string `gorm:"type:varchar(256)" json:"do_do_id" binding:"required"`
 	CFXAddress string `gorm:"type:varchar(256)" json:"cfx_address" binding:"required"`
 }
 
 type CustomMintCount struct {
 	BaseModel
 	ChannelId string `gorm:"type:varchar(256)" json:"channel_id" binding:"required"`
-	UserId string `gorm:"type:varchar(256)" json:"user_id" binding:"required"`
-	Count uint `gorm:"type:integer" json:"count" binding:"required"`
+	UserId    string `gorm:"type:varchar(256)" json:"user_id" binding:"required"`
+	Count     uint   `gorm:"type:integer" json:"count" binding:"required"`
 }
 
 func FindDiscordBindingCFXAddressById(id string) (*BindCFXWithDiscord, error) {
@@ -33,7 +33,7 @@ func FindDoDoBindingCFXAddressById(id string) (*BindCFXWithDoDo, error) {
 	return &item, err
 }
 
-func CheckDiscordCustomCount(id, channelId string, maxCount uint) (bool, error){
+func CheckDiscordCustomCount(id, channelId string, maxCount int32) (bool, error) {
 	config, err := FindDiscordCustomActivityConfigByChannelId(channelId)
 	if err != nil {
 		return false, err
@@ -44,18 +44,18 @@ func CheckDiscordCustomCount(id, channelId string, maxCount uint) (bool, error){
 	var item CustomMintCount
 	err = db.Where("user_id = ?", id).First(&item).Where("channel_id = ?", channelId).First(&item).Error
 	if err != nil {
-		err := InsertCustomCount(id,channelId)
+		err := InsertCustomCount(id, channelId)
 		if err != nil {
 			return false, err
 		}
 	}
-	if item.Count == maxCount {
+	if item.Count == uint(maxCount) {
 		return false, nil
 	}
 	return true, nil
 }
 
-func CheckDoDoCustomCount(id, channelId string, maxCount uint) (bool, error){
+func CheckDoDoCustomCount(id, channelId string, maxCount int32) (bool, error) {
 	config, err := FindDoDoCustomActivityConfigByChannelId(channelId)
 	if err != nil {
 		return false, err
@@ -66,18 +66,18 @@ func CheckDoDoCustomCount(id, channelId string, maxCount uint) (bool, error){
 	var item CustomMintCount
 	err = db.Where("user_id = ?", id).First(&item).Where("channel_id = ?", channelId).First(&item).Error
 	if err != nil {
-		err := InsertCustomCount(id,channelId)
+		err := InsertCustomCount(id, channelId)
 		if err != nil {
 			return false, err
 		}
 	}
-	if item.Count == maxCount {
+	if item.Count == uint(maxCount) {
 		return false, nil
 	}
 	return true, nil
 }
 
-func UpdateDiscordCustomCount(id, channelId string) (*CustomMintCount, error){
+func UpdateDiscordCustomCount(id, channelId string) (*CustomMintCount, error) {
 	var item CustomMintCount
 	err := db.Where("user_id = ?", id).First(&item).Where("channel_id = ?", channelId).First(&item).Error
 	if err != nil {
@@ -90,12 +90,12 @@ func UpdateDiscordCustomCount(id, channelId string) (*CustomMintCount, error){
 	if err != nil {
 		return nil, err
 	}
-	db.Model(&t).Update("amount", t.Amount - 1)
+	db.Model(&t).Update("amount", t.Amount-1)
 
 	return &item, nil
 }
 
-func UpdateDoDoCustomCount(id, channelId string) (*CustomMintCount, error){
+func UpdateDoDoCustomCount(id, channelId string) (*CustomMintCount, error) {
 	var item CustomMintCount
 	err := db.Where("user_id = ?", id).First(&item).Where("channel_id = ?", channelId).First(&item).Error
 	if err != nil {
@@ -108,22 +108,19 @@ func UpdateDoDoCustomCount(id, channelId string) (*CustomMintCount, error){
 	if err != nil {
 		return nil, err
 	}
-	db.Model(&t).Update("amount", t.Amount - 1)
+	db.Model(&t).Update("amount", t.Amount-1)
 
 	return &item, nil
 }
 
-func InsertCustomCount(id, channelId string) error{
+func InsertCustomCount(id, channelId string) error {
 	res := db.Create(&CustomMintCount{
 		ChannelId: channelId,
-		UserId: id,
-		Count: 0,
+		UserId:    id,
+		Count:     0,
 	})
 	if res.Error != nil {
-		return  res.Error
+		return res.Error
 	}
 	return nil
 }
-
-
-
